Fix to ignore down replicas

+++ ambry-coordinator/src/test/java/com.github.ambry.coordinator/CoordinatorTest.java
@@ -892,24 +892,29 @@ public class CoordinatorTest {
       throws JSONException, InterruptedException, StoreException, IOException, CoordinatorException {
 
     for (ServerErrorCode deleteErrorCode : deleteErrorMappings.keySet()) {
-      MockConnectionPool.mockCluster = new MockCluster(clusterMap);
-      induceDeleteFailure(TOTAL_HOST_COUNT, deleteErrorCode);
-      AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
-      for (int i = 0; i < 20; ++i) {
-        BlobProperties putBlobProperties =
-            new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
-        ByteBuffer putUserMetadata = getByteBuffer(10, false);
-        ByteBuffer putContent = getByteBuffer(100, true);
-        String blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
-        getBlob(ac, blobId, putBlobProperties, putUserMetadata, putContent);
-        try {
-          deleteBlobAndGet(ac, blobId);
-          fail("Deletion should have failed for " + deleteErrorCode);
-        } catch (CoordinatorException e) {
-          assertEquals(e.getErrorCode(), deleteErrorMappings.get(deleteErrorCode));
+      if (!deleteErrorCode.equals(ServerErrorCode.Disk_Unavailable) &&
+          !deleteErrorCode.equals(ServerErrorCode.Partition_ReadOnly) &&
+          !deleteErrorCode.equals(ServerErrorCode.IO_Error)) {
+        // Ignoring three ServerErrorCodes as PUTs are expected to fail in such cases
+        MockConnectionPool.mockCluster = new MockCluster(clusterMap);
+        induceDeleteFailure(TOTAL_HOST_COUNT, deleteErrorCode);
+        AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
+        for (int i = 0; i < 20; ++i) {
+          BlobProperties putBlobProperties =
+              new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
+          ByteBuffer putUserMetadata = getByteBuffer(10, false);
+          ByteBuffer putContent = getByteBuffer(100, true);
+          String blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
+          getBlob(ac, blobId, putBlobProperties, putUserMetadata, putContent);
+          try {
+            deleteBlobAndGet(ac, blobId);
+            fail("Deletion should have failed for " + deleteErrorCode);
+          } catch (CoordinatorException e) {
+            assertEquals(e.getErrorCode(), deleteErrorMappings.get(deleteErrorCode));
+          }
         }
+        ac.close();
       }
-      ac.close();
     }
   }
 
@@ -942,23 +947,59 @@ public class CoordinatorTest {
       throws JSONException, InterruptedException, StoreException, IOException, CoordinatorException {
 
     for (ServerErrorCode putErrorCode : putErrorMappings.keySet()) {
-      MockConnectionPool.mockCluster = new MockCluster(clusterMap);
-      inducePutFailure(TOTAL_HOST_COUNT, putErrorCode);
-      AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
-      for (int i = 0; i < 20; ++i) {
-        BlobProperties putBlobProperties =
-            new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
-        ByteBuffer putUserMetadata = getByteBuffer(10, false);
-        ByteBuffer putContent = getByteBuffer(100, true);
-        String blobId = null;
-        try {
-          blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
-          fail("Put should have failed for " + putErrorCode);
-        } catch (CoordinatorException e) {
-          assertEquals(e.getErrorCode(), putErrorMappings.get(putErrorCode));
+      System.out.println("puError code " + putErrorCode);
+      if (!putErrorCode.equals(ServerErrorCode.Disk_Unavailable) &&
+          !putErrorCode.equals(ServerErrorCode.Partition_ReadOnly) &&
+          !putErrorCode.equals(ServerErrorCode.IO_Error)) {
+        // Ignoring three ServerErrorCodes as PUTs are expected to fail in such cases
+        MockConnectionPool.mockCluster = new MockCluster(clusterMap);
+        inducePutFailure(TOTAL_HOST_COUNT, putErrorCode);
+        AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
+        for (int i = 0; i < 20; ++i) {
+          BlobProperties putBlobProperties =
+              new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
+          ByteBuffer putUserMetadata = getByteBuffer(10, false);
+          ByteBuffer putContent = getByteBuffer(100, true);
+          String blobId = null;
+          try {
+            blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
+            fail("Put should have failed for " + putErrorCode);
+          } catch (CoordinatorException e) {
+            assertEquals(e.getErrorCode(), putErrorMappings.get(putErrorCode));
+          }
         }
+        ac.close();
       }
-      ac.close();
+    }
+  }
+
+  void putExceptionOnUnavailability(ClusterMap clusterMap)
+      throws JSONException, InterruptedException, StoreException, IOException, CoordinatorException {
+
+    for (ServerErrorCode putErrorCode : putErrorMappings.keySet()) {
+      if (putErrorCode.equals(ServerErrorCode.Disk_Unavailable) ||
+          putErrorCode.equals(ServerErrorCode.Partition_ReadOnly) ||
+          putErrorCode.equals(ServerErrorCode.IO_Error)) {
+        // Considering only those error codes which would cause the replica to be down which in turn will affect the
+        // replicas to be considered for put policy
+        MockConnectionPool.mockCluster = new MockCluster(clusterMap);
+        inducePutFailure(TOTAL_HOST_COUNT, putErrorCode);
+        AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
+        for (int i = 0; i < 20; ++i) {
+          BlobProperties putBlobProperties =
+              new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
+          ByteBuffer putUserMetadata = getByteBuffer(10, false);
+          ByteBuffer putContent = getByteBuffer(100, true);
+          String blobId = null;
+          try {
+            blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
+            fail("Put should have failed for " + putErrorCode);
+          } catch (CoordinatorException e) {
+            assertEquals(e.getErrorCode(), CoordinatorError.UnexpectedInternalError);
+          }
+        }
+        ac.close();
+      }
     }
   }
 
@@ -999,6 +1040,12 @@ public class CoordinatorTest {
   public void simpleOneDCThreeNodeOneDiskOnePartitionForPutException()
       throws JSONException, InterruptedException, StoreException, IOException, CoordinatorException {
     simplePutException(getClusterMapOneDCThreeNodeOneDiskOnePartition());
+  }
+
+  @Test
+  public void simpleOneDCThreeNodeOneDiskOnePartitionForPutExceptionOnUnavailability()
+      throws JSONException, InterruptedException, StoreException, IOException, CoordinatorException {
+    putExceptionOnUnavailability(getClusterMapOneDCThreeNodeOneDiskOnePartition());
   }
 
   @Test
